refactor(config): extract user bindings
